Refactor context files to fix React Fast Refresh warnings
