+++ src/context/AuthContext.jsx
@@ -1,4 +1,4 @@
-/* eslint-disable react-refresh/only-export-components */
+
 import React, {
   createContext,
   useContext,
@@ -196,7 +196,7 @@ export const AuthProvider = ({ children }) => {
       // Set user immediately and stop loading - don't wait for Firestore
       setCurrentUser(user);
       setLoading(false);
-      
+
       if (user) {
         // Fetch additional user data from Firestore in the background
         try {
@@ -281,4 +281,4 @@ export const AuthProvider = ({ children }) => {
   );
 };
 
-export default AuthContext;
+
