+++ src/context/ThemeContext.jsx
@@ -1,4 +1,4 @@
-/* eslint-disable react-refresh/only-export-components */
+
 import React, { createContext, useContext, useState, useEffect, useMemo, useCallback } from "react";
 
 const ThemeContext = createContext();
@@ -42,4 +42,4 @@ export const ThemeProvider = ({ children }) => {
   );
 };
 
-export default ThemeContext;
+
